# Optimization for `cache.parsePath` to reduce allocations